pyproject.toml
[project]
name = "reco-system"
version = "0.1.0"
description = ""
authors = [
    {name = "LucBourrat1",email = "luc.bourrat@gmail.com"}
]
readme = "README.md"
requires-python = "3.10.12"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "numpy (>=1.24.0,<2.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "scikit-surprise",
    "flask (>=3.1.0,<4.0.0)"
]

[tool.poetry]
packages = [{include = "reco_system", from = "src"}]


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
